Split service.rs into host and device submodules
