+++ rust-remote-shell/src/lib.rs
@@ -1,4 +1,6 @@
+pub mod device;
 pub mod device_server;
+pub mod host;
 pub mod io_handler;
 pub mod sender_client;
 pub mod service;
